# chore: simplify code generation logic for a faster and better UX