--- a/storage/pipeline/fsm.go
+++ b/storage/pipeline/fsm.go
@@ -1,5 +1,3 @@
-//go:generate go run ./gen
-
 package sealing
 
 import (