## Related Issues
* Fixes #8392

## Proposed Changes
The changes here simplify the string of targets executed as part of `make gen`. The simplifications include:
* parallelised code generation whenever possible.
* re-implementation of documentation generation for a significant simplification and improved readability.
* unified mocks generation to avoid multiple slow calls to `go run` per package.

Note, the changes introduced here are purely mechanical and do not alter the Lotus node functionality.

## Additional Info

### Timing On CI

* Before: [2m 41s](https://github.com/filecoin-project/lotus/actions/runs/10938947953/job/30368161181)
* After: [1m 28s](https://github.com/filecoin-project/lotus/actions/runs/10943431852/job/30382837038?pr=12493)

### Timing On my machine
* Before: `make gen  105.22s user 56.84s system 152% cpu 1:46.46 total`
* After: `make gen  54.30s user 36.99s system 178% cpu 51.247 total`

## Checklist

Before you mark the PR ready for review, please make sure that:

- [x] Commits have a clear commit message.
- [x] PR title conforms with [contribution conventions](https://github.com/filecoin-project/lotus/blob/master/CONTRIBUTING.md#pr-title-conventions)
- [x] Update CHANGELOG.md or signal that this change does not need it per [contribution conventions](https://github.com/filecoin-project/lotus/blob/master/CONTRIBUTING.md#changelog-management)
- [ ] New features have usage guidelines and / or documentation updates in
  - [ ] [Lotus Documentation](https://lotus.filecoin.io)
  - [ ] [Discussion Tutorials](https://github.com/filecoin-project/lotus/discussions/categories/tutorials)
- [x] Tests exist for new functionality or change in behavior
- [x] CI is green