@rjan90 2024-09-19
I still have some cleanups/updates to do on my computer that makes it a little slow. But this is soooo much faster then when I ran a couple of days ago:

`make gen  74.23s user 22.34s system 127% cpu 1:15.79 total`