@masih 2024-09-19
There is room for more improvements, e.g. not generating files that need fiximports, etc. but the correct fix IMHO is to simply do a lot less. 

For now, more CPU cores to the rescue.